fastapi_v1/requirements.txt
bcrypt==3.2.0
scikit-learn==0.23.2
fastapi==0.61.1
uvicorn==0.12.1
pandas==1.1.2
Keras==2.3.1
pydantic==1.6.1
numpy==1.19.2
pickle5==0.0.11
cloudpickle==1.6.0
tensorflow==2.2.0
scipy==1.4.1
passlib==1.7.4
cryptography==2.9.2
python-jose[cryptography]==3.2.0
python-multipart==0.0.5

